+++ src/utils/basket.ts
@@ -1,6 +1,6 @@
 export const calculateBasketTotal = (items: Product[]) => {
   return items.reduce(
-    (acc, item) => acc + parseInt(item.unitPrice) * parseInt(item.quantity),
+    (acc, item) => acc + parseFloat(item.unitPrice) * parseFloat(item.quantity),
     0
   )
 }
